third_party/blink/web_tests/platform/linux-chrome/external/wpt/html-aam/fragile/optgroup-role-expected.txt

This is a testharness.js-based test.
[FAIL] el-optgroup
  assert_equals: <optgroup label="x" data-testname="el-optgroup" data-expectedrole="group" class="ex">\n    <option>x1</option>\n    <option>x2</option>\n  </optgroup> expected "group" but got "none"
Harness: the test ran to completion.
